requirements.txt
lancedb
pyarrow
tornado
pyjwt
google-generativeai
asyncio
numpy
pandas